Update textual to 0.24.0 and fix stuff
The Autocomplete widget interferes with the labeled containers that we
use starting from this version. It's probably due to the fact that the
autocomplete widget attempts to manipulate layers in a dynamic way (to
create a layer on top) and that interferes with the labeled
containers, which use layers to display the label.

This version also happened to introduce a Select widget, which we can
use as a replacement. The functionality is not exactly the same (it
requires choosing from a dropdown list without being able to search by
typing), but it will have to do for now. I've tried to replace the
widget type with the minimal diff possible.