Update `textual` dependency (#45)
* Update textual to 0.21.0 and fix create screen's scrollbars

* Fix height on sidebar of status screen so that it fits in window

* Prevent crash due to `on_click` not existing in parent class

* Prevent crash due to calls to `on_mount()`

* Update textual to 0.24.0 and fix stuff

The Autocomplete widget interferes with the labeled containers that we
use starting from this version. It's probably due to the fact that the
autocomplete widget attempts to manipulate layers in a dynamic way (to
create a layer on top) and that interferes with the labeled
containers, which use layers to display the label.

This version also happened to introduce a Select widget, which we can
use as a replacement. The functionality is not exactly the same (it
requires choosing from a dropdown list without being able to search by
typing), but it will have to do for now. I've tried to replace the
widget type with the minimal diff possible.

* Update to latest Textual and make necessary adjustments

This update surfaced exceptions that occur when the config has fields
with incorrect types. That situation is problematic for populating the
configuration screen without having the Pydantic model to guarantee
the shape of the data.

* Satisfy mypy

* Show scrollbars for errors